Bug fixes of delete comment and Answer comment

+++ application/routes.py
@@ -1,8 +1,8 @@
 import requests as req
 from flask import render_template, url_for, flash, redirect, request, abort, jsonify
 from application import app, db, bcrypt, mail
-from application.forms import RegistrationForm, LoginForm, UpdateAccountForm, PostForm, CommentForm, RequestResetForm, ResetPasswordForm
-from application.models import User, Post, Answer, Comment, Vote, Downvote, Tags, tagposts, Answerupvotes, Answerdownvotes, Favourite
+from application.forms import RegistrationForm, LoginForm, UpdateAccountForm, PostForm, CommentForm, RequestResetForm, ResetPasswordForm, AnswerCommentForm
+from application.models import User, Post, Answer, Comment, Vote, Downvote, Tags, tagposts, Answerupvotes, Answerdownvotes, Favourite, AnswerComment
 from flask_login import login_user, current_user, logout_user, login_required
 import secrets
 import os
@@ -14,7 +14,6 @@ from sqlalchemy import and_, or_
 from sqlalchemy import create_engine
 import json
 from flask_mail import Message
-
 
 
 def dict_factory(cursor, row):
@@ -42,16 +41,16 @@ def create_connection(db_file):
 def home():
     page = request.args.get('page',1,type=int)
     posts = Post.query.order_by(Post.date_posted.desc()).paginate(page=page, per_page=5)
-    
+
 
     # Join tables (tagposts, Tags and Post) to get tags for each post
     joinedTables = db.session.query(tagposts.post_id,tagposts.tag_id,Tags.tag_title,Post.title,Post.user_id,Post.like_count,Post.content).join(Tags).join(Post).all()
-    
+
     recommended_items_default = []
     items = db.session.query(Post).order_by(Post.date_posted.desc()).all()
     if not items:
         recommended_items_default.clear()
-    
+
     else:
         for x in items:
             recommended_items_default.append(x.id)
@@ -68,12 +67,12 @@ def home():
         if str(response) == '<Response [200]>':
             print(response.text)
             print('resp 200')
-            
+
             recommended_items = json.loads(response.text)
             if not recommended_items:
                 flash("No posts on the site, Add a post to start!", 'info')
                 return redirect('/post/new')
-               
+
             else:
                 return render_template('home.html', posts=posts, joinedTables=joinedTables, recommended_items=recommended_items)
         else:
@@ -88,13 +87,13 @@ def home():
 def personal():
     page = request.args.get('page',1,type=int)
     posts = Post.query.order_by(Post.date_posted.desc()).paginate(page=page, per_page=5)
-    
+
     # Join tables (tagposts, Tags and Post) to get tags for each post
     joinedTables = db.session.query(tagposts.post_id,tagposts.tag_id,Tags.tag_title,Post.title,Post.user_id,Post.like_count,Post.content).join(Tags).join(Post).all()
 
     if current_user.is_authenticated:
         uid = current_user.id
-        
+
         # API call
         furl = "http://mayankms02.pythonanywhere.com/api/sCLAzIJDKmi1SfWLeapXYA/data_delivery/tag_item"
         params = {"uid": uid}
@@ -112,11 +111,11 @@ def personal():
                 int_recommended_list.append(int(x))
             print(int_recommended_list)
             for x in recommended_items:
-                
+
                 post = db.session.query(Post).filter_by(id=x).all()
                 print(x)
                 print(post)
-        
+
 
                 # post = db.session.query(Post).all()
                 return render_template('personal.html', posts=posts, joinedTables=joinedTables, post=post, recommended_items=int_recommended_list)
@@ -278,7 +277,7 @@ def new_post():
 
         for one_tag in many_tags:
             # print(one_tag)
-           
+
             # check if tag is  present in table
             tag_present = db.session.query(db.exists().where(and_(Tags.tag_title == one_tag))).scalar()
             if tag_present:
@@ -317,7 +316,7 @@ def new_post():
         else:
             print('api call error')
             return redirect(url_for('home'))
-        
+
     elif request.method == 'GET':
         dbase = create_connection("site.db")
         sql = 'select * from Tags'
@@ -348,32 +347,38 @@ def update_post(post_id):
     post_dict['post_title'] = post.title
     post_dict['post_content'] = post.content
 
-
-    # If post author is not current user then abort
-    if post.author != current_user:
-        abort(403)
-    # Else if current user is author then import the form
     if request.method == 'POST':
         post.title = request.form.get('title')
         post.content = request.form.get('content')
         db.session.commit()
+        taglist = []
 
         # Get all tags on the post
         many_tags = request.form.getlist('tags')
         for one_tag in many_tags:
-            print(one_tag)
-            tag = Tags(tag_title=one_tag)
-            db.session.add(tag)
-            db.session.commit()
-        flash('Your question has been updated', 'success')
+            # check if tag is  present in table
+            tag_present = db.session.query(db.exists().where(and_(Tags.tag_title == one_tag))).scalar()
+            if tag_present:
+                tag_row = db.session.query(Tags).filter_by(tag_title=one_tag).all()
+                # tagid of old tags in a taglist
+                for item in tag_row:
+                    taglist.append(item.tag_id)
+            else:
+                tag = Tags(tag_title=one_tag)
+                db.session.add(tag)
+                db.session.commit()
+                tag_row = db.session.query(Tags).filter_by(tag_title=one_tag).all()
+                for item in tag_row:
+                    taglist.append(item.tag_id)
+        print("taglist: ", taglist)
 
-        for tag in recent_tags: 
-            print(tag)
-            tag_id = tag.tag_id
-            something = tagposts(post_id=post_id, tag_id=tag_id)
+        for tag in taglist:
+            something = tagposts(post_id=post_id, tag_id=tag)
             db.session.add(something)
             db.session.commit()
-        return redirect(url_for('home'))
+
+
+            return redirect(url_for('comment_post', post_id=post_id))
 
     # Populate form with current post data
     elif request.method == 'GET':
@@ -384,7 +389,7 @@ def update_post(post_id):
             taglist = dbase.execute(sql).fetchall()
         except Exception as e:
             print(e)
-      
+
         return render_template('update.html', post=post_dict, joinedTables=joinedTables, title='Update Post', legend='Edit Your Question', taglist=taglist)
 
 # ---------- Delete Post route ---------
@@ -429,7 +434,7 @@ def user_posts(username):
 def comment_post(post_id):
     post = Post.query.get_or_404(post_id)
     posts_id = post.id
-
+    
     # Join tables (tagposts, Tags and Post) to get tags for each post
     joinedTables = db.session.query(tagposts.post_id,tagposts.tag_id,Tags.tag_title,Post.title,Post.user_id,Post.like_count,Post.content).join(Tags).join(Post).all()
 
@@ -439,26 +444,28 @@ def comment_post(post_id):
     # Get all answers from Answer table
     answers = Answer.query.order_by(Answer.date_posted.desc()).all()
 
-    form = CommentForm()
-    if form.validate_on_submit():
-        comment = Comment(comment=form.comment.data, post_id=posts_id, user_id=current_user.id)
+    if request.method == 'POST':
+        comment = Comment(comment=request.form.get('comment'), post_id=posts_id, user_id=current_user.id)
         post = Post.query.get_or_404(post_id)
         db.session.add(comment)
         db.session.commit()
         print(current_user.id)
         comments = Comment.query.filter(Comment.post_id == post_id).order_by(Comment.date_posted.desc())
-        # flash('Your comment has been posted', 'success')
-        return render_template('comment_post.html', title='Comments', form=form, post=post, comments=comments, joinedTables=joinedTables, answersCount=answersCount, answers=answers)
-    else:
+        flash('Your comment has been added', 'success')
+
+        return render_template('comment_post.html', title='Comments', post=post, comments=comments, joinedTables=joinedTables, answersCount=answersCount, answers=answers)
+
+    if request.method =='GET':
         comments = Comment.query.filter(Comment.post_id == post_id).order_by(Comment.date_posted.desc())
-        return render_template('comment_post.html', title='Comments', form=form, post=post, comments=comments, joinedTables=joinedTables, answersCount=answersCount, answers=answers)
+        return render_template('comment_post.html', title='Comments', post=post, comments=comments, joinedTables=joinedTables, answersCount=answersCount, answers=answers)
+
 
 # ---------- Search route ---------
 @app.route("/search", methods=['GET', 'POST'])
 def search():
         page = request.args.get('page', 1, type=int)
         searched_content = request.form.get('search')
-        
+
         searched_posts = Post.query.filter(or_(Post.title.like('%'+searched_content+'%'),Post.content.like('%'+searched_content+'%'))).paginate(page=page, per_page=5)
 
         return render_template('search.html', searched_posts=searched_posts)
@@ -474,7 +481,7 @@ def upvote(post_id,user_id):
 
     # Check if user has Disliked this post or not
     post_inDownvote = db.session.query(db.exists().where(and_(Downvote.post_id == post_id, Downvote.user_id == user_id))).scalar()
-    
+
 
     if post_present == True:
         print("All ready upvoted")
@@ -485,7 +492,7 @@ def upvote(post_id,user_id):
         # If Yes, Delete the entry
         Vote.query.filter_by(post_id=post_id, user_id=user_id).delete()
         db.session.commit()
-        
+
         # API call
         furl = "https://mayankms02.pythonanywhere.com/api/sCLAzIJDKmi1SfWLeapXYA/data_entry/user_action"
         params = {"uid": user_id, "itemid": post_id, "action": 0}
@@ -497,7 +504,7 @@ def upvote(post_id,user_id):
         upvoteCount = db.session.query(Vote).filter(Vote.post_id == post_id).count()
         posts.like_count = upvoteCount
         db.session.commit()
-        
+
         print(posts.like_count)
         print(posts)
         print("inside if")
@@ -520,7 +527,7 @@ def upvote(post_id,user_id):
         vote.action = 1
         db.session.add(vote)
         db.session.commit()
-        
+
         # API call
         furl = "https://mayankms02.pythonanywhere.com/api/sCLAzIJDKmi1SfWLeapXYA/data_entry/user_action"
         params = {"uid": user_id, "itemid": post_id, "action": 1}
@@ -553,7 +560,7 @@ def upvote(post_id,user_id):
         vote.action = 1
         db.session.add(vote)
         db.session.commit()
-        
+
         # API call
         furl = "https://mayankms02.pythonanywhere.com/api/sCLAzIJDKmi1SfWLeapXYA/data_entry/user_action"
         params = {"uid": user_id, "itemid": post_id, "action": 1}
@@ -576,7 +583,7 @@ def upvote(post_id,user_id):
         vote_data = {'vote_count': upvote, 'downvote_count': downvote}
         print('vote_data', vote_data)
         return jsonify(vote_data)
-        
+
 
 # ---------- Downvote route ---------
 @app.route('/downvote/<int:post_id>/<int:user_id>')
@@ -593,11 +600,11 @@ def downvote(post_id,user_id):
     if post_present == True:
         print("All ready downvoted")
         print(post_present)
-        
+
         # If YES, Delete the entry
         Downvote.query.filter_by(post_id=post_id, user_id=user_id).delete()
         db.session.commit()
-        
+
         # API call
         furl = "https://mayankms02.pythonanywhere.com/api/sCLAzIJDKmi1SfWLeapXYA/data_entry/user_action"
         params = {"uid": user_id, "itemid": post_id, "action": 0}
@@ -609,7 +616,7 @@ def downvote(post_id,user_id):
         DownvoteCount = db.session.query(Downvote).filter(Downvote.post_id == post_id).count()
         posts.dislike_count = DownvoteCount
         db.session.commit()
-        
+
         upvote = db.session.query(Vote).filter(Vote.post_id == post_id).count()
         downvote = db.session.query(Downvote).filter(Downvote.post_id == post_id).count()
 
@@ -628,7 +635,7 @@ def downvote(post_id,user_id):
         downvote.action = "disliked"
         db.session.add(downvote)
         db.session.commit()
-        
+
 
         # API call
         furl = "https://mayankms02.pythonanywhere.com/api/sCLAzIJDKmi1SfWLeapXYA/data_entry/user_action"
@@ -681,7 +688,7 @@ def downvote(post_id,user_id):
         vote_data = {'vote_count': upvote, 'downvote_count': downvote}
         print('vote_data', vote_data)
         return jsonify(vote_data)
-    
+
 
 # ---------- Tags route ---------
 @app.route('/tags/<tag_title>')
@@ -697,24 +704,30 @@ def tags(tag_title):
     for item in joinedTables:
         if tag_title in item.tag_title:
             print(tag_title, item.post_id, item.title, item.user_id,  item.username, item.image_file)
-            
+
 
             return render_template("tags.html", posts=posts, joinedTables=joinedTables, tag_title=tag_title)
 
 # ---------- Delete Comment route ---------
-@app.route("/comment/<int:post_id>/<int:comment_id>/delete", methods=['GET','POST'])
+@app.route("/post/<post_id>/comment/<comment_id>/delete", methods=['POST'])
 @login_required
 def delete_comment(post_id,comment_id):
-    comment = Comment.query.get_or_404(comment_id)
-    print(comment)
-    post = Post.query.get_or_404(post_id)
-    print(post)
-    if post.author != current_user:
-        abort(403)
-    db.session.delete(comment)
-    db.session.commit()
-    flash('Comment has been deleted', 'success')
-    return redirect(url_for('comment_post', post_id=post_id))
+    if request.method == 'POST':
+        
+        comment_exists = db.session.query(db.exists().where(and_(Comment.post_id == post_id, Comment.comment_id == comment_id))).scalar()
+        print(comment_exists)
+        if comment_exists:
+            comment = Comment.query.get(comment_id)
+            print(comment)
+            post = Post.query.get(post_id)
+            print(post)
+            if post.author != current_user:
+                return jsonify(False)
+            db.session.delete(comment)
+            db.session.commit()
+            # flash('Comment has been deleted', 'success')
+            return redirect(url_for('comment_post', post_id=post_id))
+        # return jsonify('check logs')
 
 # ---------- Answer route ---------
 @app.route("/answer/<int:post_id>", methods=['GET', 'POST'])
@@ -724,13 +737,35 @@ def answer(post_id):
         answer = Answer(content=request.form.get('editordata'), author=current_user, post_id=post_id)
         db.session.add(answer)
         db.session.commit()
-    
+
         post = Post.query.get_or_404(post_id)
         posts_id = post.id
-        
+
         return redirect(url_for('comment_post', post_id=post_id))
 
     return redirect(url_for('comment_post', post_id=post_id))
+
+# ---------- Answer Comment route ---------
+@app.route("/answer/<int:answer_id>/comment", methods=['GET', 'POST'])
+@login_required
+def answercomment(answer_id):
+    answer = Answer.query.get_or_404(answer_id)
+    answer_id = answer.id
+    print(answer)
+
+    form = AnswerCommentForm()
+    if form.validate_on_submit():
+        comment = AnswerComment(comment=form.comment.data, answer_id=answer_id, user_id=current_user.id)
+        answer = Answer.query.get_or_404(answer_id)
+        db.session.add(comment)
+        db.session.commit()
+        comments = AnswerComment.query.filter(AnswerComment.answer_id == answer_id).all()
+
+        return render_template('answer_comment.html', title='Comments', form=form, comments=comments, answer=answer)
+    else:
+        comments = AnswerComment.query.filter(AnswerComment.answer_id == answer_id).all()
+        return render_template('answer_comment.html', title='Comments', form=form, comments=comments, answer=answer)
+
 
 # ---------- Delete Answer route ---------
 @app.route("/answer/<int:answer_id>/delete", methods=['POST'])
@@ -772,7 +807,7 @@ def answerLikes(answer_id,user_id):
     answer_present = db.session.query(db.exists().where(and_(Answerupvotes.answer_id == answer_id, Answerupvotes.user_id == user_id))).scalar()
     # Check if user has Disliked this post or not
     post_inAnswerDownvote = db.session.query(db.exists().where(and_(Answerdownvotes.answer_id == answer_id, Answerdownvotes.user_id == user_id))).scalar()
-    
+
 
     if answer_present == True:
         print("All ready upvoted")
@@ -858,7 +893,7 @@ def answerDislikes(answer_id,user_id):
     answer_present = db.session.query(db.exists().where(and_(Answerdownvotes.answer_id == answer_id, Answerdownvotes.user_id == user_id))).scalar()
     # Check if user has Liked this post or not
     post_inAnswerUpvote = db.session.query(db.exists().where(and_(Answerupvotes.answer_id == answer_id, Answerupvotes.user_id == user_id))).scalar()
-    
+
 
     if answer_present == True:
         print("All ready Downvoted")
@@ -952,11 +987,11 @@ def update_answer(answer_id, post_id):
         answer.content = request.form.get('editordata')
         db.session.commit()
         return redirect(url_for('comment_post', post_id=post_id))
-        
+
 
     # Populate form with current answer data
     elif request.method == 'GET':
-      
+
         return render_template('editAnswer.html',post=post, answer=answer, joinedTables=joinedTables, title='Update Answer', legend='Edit Your Answer')
 
 # ---------- Favourite Posts routes ----------
@@ -973,7 +1008,7 @@ def favourite_post(post_id):
         db.session.add(addFavourite)
         db.session.commit()
         flash("Post has been added to favourites", 'success')
-    
+
     return redirect(url_for('home'))
 
 # ---------- Show Favourite Posts routes ----------
@@ -988,7 +1023,7 @@ def favourite():
     # print(joinedTables)
 
 
-            
+
 
     return render_template("favourite.html", posts=posts, joinedTables=joinedTables)
